# generated/java/src/main/java/saas/identity/shared/api/TenantAuditApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package saas.identity.shared.api;

import saas.identity.shared.dto.AuditAction;
import org.springframework.format.annotation.DateTimeFormat;
import saas.identity.shared.dto.ErrorResponse;
import org.springframework.lang.Nullable;
import java.time.OffsetDateTime;
import saas.identity.shared.dto.TenantAuditExportAuditEvents200Response;
import saas.identity.shared.dto.TenantAuditExportAuditEventsRequest;
import saas.identity.shared.dto.TenantAuditGetRetentionPolicy200Response;
import saas.identity.shared.dto.TenantAuditListAuditEvents200Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-12T12:50:18.473810800+08:00[Asia/Shanghai]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "tenant-audit", description = "the tenant-audit API")
public interface TenantAuditApi {

    String PATH_TENANT_AUDIT_EXPORT_AUDIT_EVENTS = "/api/v1/tenants/{tenantId}/audit-events/export";
    /**
     * POST /api/v1/tenants/{tenantId}/audit-events/export
     *
     * @param tenantId  (required)
     * @param tenantAuditExportAuditEventsRequest  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantAuditExportAuditEvents",
        tags = { "tenant-audit" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantAuditExportAuditEvents200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TenantAuditApi.PATH_TENANT_AUDIT_EXPORT_AUDIT_EVENTS,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TenantAuditExportAuditEvents200Response> tenantAuditExportAuditEvents(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "TenantAuditExportAuditEventsRequest", description = "", required = true) @Valid @RequestBody TenantAuditExportAuditEventsRequest tenantAuditExportAuditEventsRequest
    );


    String PATH_TENANT_AUDIT_GET_RETENTION_POLICY = "/api/v1/tenants/{tenantId}/audit-events/retention";
    /**
     * GET /api/v1/tenants/{tenantId}/audit-events/retention
     *
     * @param tenantId  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantAuditGetRetentionPolicy",
        tags = { "tenant-audit" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantAuditGetRetentionPolicy200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TenantAuditApi.PATH_TENANT_AUDIT_GET_RETENTION_POLICY,
        produces = { "application/json" }
    )
    ResponseEntity<TenantAuditGetRetentionPolicy200Response> tenantAuditGetRetentionPolicy(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId
    );


    String PATH_TENANT_AUDIT_LIST_AUDIT_EVENTS = "/api/v1/tenants/{tenantId}/audit-events";
    /**
     * GET /api/v1/tenants/{tenantId}/audit-events
     *
     * @param tenantId  (required)
     * @param page  (optional)
     * @param pageSize  (optional)
     * @param actorUserId  (optional)
     * @param action  (optional)
     * @param from  (optional)
     * @param to  (optional)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantAuditListAuditEvents",
        tags = { "tenant-audit" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantAuditListAuditEvents200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TenantAuditApi.PATH_TENANT_AUDIT_LIST_AUDIT_EVENTS,
        produces = { "application/json" }
    )
    ResponseEntity<TenantAuditListAuditEvents200Response> tenantAuditListAuditEvents(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) @Nullable Integer page,
        @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false) @Nullable Integer pageSize,
        @Parameter(name = "actorUserId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "actorUserId", required = false) @Nullable String actorUserId,
        @Parameter(name = "action", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "action", required = false) @Nullable AuditAction action,
        @Parameter(name = "from", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "from", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) @Nullable OffsetDateTime from,
        @Parameter(name = "to", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "to", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) @Nullable OffsetDateTime to
    );


    String PATH_TENANT_AUDIT_LIST_AUDIT_EVENTS_BY_USER = "/api/v1/tenants/{tenantId}/audit-events/by-user/{userId}";
    /**
     * GET /api/v1/tenants/{tenantId}/audit-events/by-user/{userId}
     *
     * @param tenantId  (required)
     * @param userId  (required)
     * @param page  (optional)
     * @param pageSize  (optional)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantAuditListAuditEventsByUser",
        tags = { "tenant-audit" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantAuditListAuditEvents200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TenantAuditApi.PATH_TENANT_AUDIT_LIST_AUDIT_EVENTS_BY_USER,
        produces = { "application/json" }
    )
    ResponseEntity<TenantAuditListAuditEvents200Response> tenantAuditListAuditEventsByUser(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") String userId,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) @Nullable Integer page,
        @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false) @Nullable Integer pageSize
    );


    String PATH_TENANT_AUDIT_SET_RETENTION_POLICY = "/api/v1/tenants/{tenantId}/audit-events/retention";
    /**
     * PUT /api/v1/tenants/{tenantId}/audit-events/retention
     *
     * @param tenantId  (required)
     * @param tenantAuditGetRetentionPolicy200Response  (required)
     * @return The request has succeeded. (status code 200)
     *         or An unexpected error response. (status code 200)
     */
    @Operation(
        operationId = "tenantAuditSetRetentionPolicy",
        tags = { "tenant-audit" },
        responses = {
            @ApiResponse(responseCode = "200", description = "The request has succeeded.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TenantAuditGetRetentionPolicy200Response.class))
            }),
            @ApiResponse(responseCode = "default", description = "An unexpected error response.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = TenantAuditApi.PATH_TENANT_AUDIT_SET_RETENTION_POLICY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TenantAuditGetRetentionPolicy200Response> tenantAuditSetRetentionPolicy(
        @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") String tenantId,
        @Parameter(name = "TenantAuditGetRetentionPolicy200Response", description = "", required = true) @Valid @RequestBody TenantAuditGetRetentionPolicy200Response tenantAuditGetRetentionPolicy200Response
    );

}