# generated/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /api/v1/tenants/{tenantId}/audit-events:
    get:
      operationId: TenantAudit_listAuditEvents
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: actorUserId
          in: query
          required: false
          schema:
            type: string
          explode: false
        - name: action
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AuditAction'
          explode: false
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - page
                  - pageSize
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
                  page:
                    type: integer
                    format: int32
                    example: 0
                  pageSize:
                    type: integer
                    format: int32
                    example: 20
                  total:
                    type: integer
                    format: int64
                    example: 0
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-audit
  /api/v1/tenants/{tenantId}/audit-events/by-user/{userId}:
    get:
      operationId: TenantAudit_listAuditEventsByUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - page
                  - pageSize
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
                  page:
                    type: integer
                    format: int32
                    example: 0
                  pageSize:
                    type: integer
                    format: int32
                    example: 20
                  total:
                    type: integer
                    format: int64
                    example: 0
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-audit
  /api/v1/tenants/{tenantId}/audit-events/export:
    post:
      operationId: TenantAudit_exportAuditEvents
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                required:
                  - downloadUrl
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
                format:
                  type: string
                  enum:
                    - csv
                    - json
              required:
                - from
                - to
                - format
  /api/v1/tenants/{tenantId}/audit-events/retention:
    get:
      operationId: TenantAudit_getRetentionPolicy
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  retentionDays:
                    type: integer
                    format: int32
                required:
                  - retentionDays
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-audit
    put:
      operationId: TenantAudit_setRetentionPolicy
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  retentionDays:
                    type: integer
                    format: int32
                required:
                  - retentionDays
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                retentionDays:
                  type: integer
                  format: int32
              required:
                - retentionDays
...
components:
  schemas:
...
    AuditAction:
      type: string
      enum:
        - user_created
        - user_updated
        - user_deleted
        - role_assigned
        - role_revoked
        - login_success
        - login_failed
        - oauth_token_issued
        - api_key_created
        - api_key_revoked
    AuditEvent:
      type: object
      required:
        - id
        - tenantId
        - action
        - occurredAt
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        actorUserId:
          type: string
          format: uuid
        action:
          $ref: '#/components/schemas/AuditAction'
        targetUserId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: {}
        occurredAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
servers:
  - url: https://api.example.com
    description: Production
    variables: {}